Set file stream length before writing file
If the new file contents are shorter than the original file, some of the data from the original file will remain in the buffer and will pollute the new data when it gets written out.  Resetting the length tells the file stream to ignore any data in the buffer that extends beyond what the updated data will use.