Extend tip types class to encapsulate business logic
Tip types will eventually be defined by the user, but the business logic will still apply.  Some payable tips (such as credit card and room charge) are accrued on a specific business date.  Claimed, non-payable tips (e.g. cash tips) are claimed for the pay period, and not for a specific date (though a date is required, so the last day of the pay period is used).  Tips accured from working an event are considered accrued on the date of the event, so the user cannot change the date for them.